Bugfix: Regression from millisecond time
The idea was to switch from measuring time in `float seconds` to `int32_t milliseconds` for better accuracy and no float rounding errors. This requires time connstants to use milliseconds though. The merge didn't account for non-confilicting files that have used seconds as time reference.

This caused cooldowns to expire almost instantly thus seem not to work.

Fix: Rollback milliseconds change, now all time is uniformly measured in `float seconds` again.
When we really do run into issues with accuracy and rounding, we can switch to milliseconds.

+++ src/Item/ItemState.h
@@ -10,7 +10,7 @@ class ItemState
 	ItemState();
 	ItemState(uint32_t id, sf::Vector2f position, PowerupType type);
 
-	void update(int32_t dt);
+	void update(float dt);
 
 	uint32_t getId() const
 	{
